cleaning up new solution
